# static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AI Interview Platform API",
    "description": "API documentation for the AI-Powered Technical Interview Platform",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API Server"
    }
  ],
...
  "paths": {
...
    "/coding/challenges": {
      "get": {
        "tags": ["Coding"],
        "summary": "Get coding challenges",
        "description": "Get coding challenges for the authenticated user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "difficulty",
            "in": "query",
            "description": "Filter by difficulty",
            "schema": {
              "type": "string",
              "enum": ["easy", "medium", "hard"]
            }
          },
          {
            "name": "language",
            "in": "query",
            "description": "Filter by programming language",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter by status",
            "schema": {
              "type": "string",
              "enum": ["completed", "pending"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of coding challenges",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Coding"],
        "summary": "Create a coding challenge",
        "description": "Create a new coding challenge (recruiters only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["candidate_id", "title", "description", "difficulty", "language"],
                "properties": {
                  "candidate_id": {
                    "type": "integer"
                  },
                  "interview_id": {
                    "type": "integer"
                  },
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "difficulty": {
                    "type": "string",
                    "enum": ["easy", "medium", "hard"]
                  },
                  "language": {
                    "type": "string"
                  },
                  "generate_test_cases": {
                    "type": "boolean"
                  },
                  "test_cases": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Coding challenge created successfully"
          },
          "400": {
            "description": "Invalid request"
          },
          "403": {
            "description": "Unauthorized access"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}